Refactor structure of behaviors
