+++ CSharpSourceCode/Battle/AI/AgentBehavior/AgentCastingBehavior/AoETargetedCastingBehavior.cs
@@ -10,8 +10,7 @@ namespace TOW_Core.Battle.AI.AgentBehavior.AgentCastingBehavior
         public AoETargetedCastingBehavior(Agent agent, AbilityTemplate template, int abilityIndex) : base(agent, template, abilityIndex)
         {
         }
-
-
+        
         public override void Execute()
         {
             if (AbilityTemplate.AbilityTargetType == AbilityTargetType.Self)
