+++ CSharpSourceCode/Battle/AI/AgentBehavior/AgentCastingBehavior/PreserveWindsCastingBehavior.cs
@@ -9,6 +9,11 @@ namespace TOW_Core.Battle.AI.AgentBehavior.AgentCastingBehavior
         public PreserveWindsAgentCastingBehavior(Agent agent, AbilityTemplate abilityTemplate, int abilityIndex) : base(agent, abilityTemplate, abilityIndex)
         {
         }
+        
+        public override void Execute()
+        {
+            //Do nothing. I am hoping that we will add some sort of "Channeling" which allows us to restore magic over time later on.
+        }
 
         public override void Terminate()
         {
@@ -18,12 +23,7 @@ namespace TOW_Core.Battle.AI.AgentBehavior.AgentCastingBehavior
         {
             return false;
         }
-
-        public override void Execute()
-        {
-            //Do nothing. I am hoping that we will add some sort of "Channeling" which allows us to restore magic over time later on.
-        }
-
+        
         protected override Target CalculateUtility(Target target)
         {
             target.UtilityValue = 0.4f;
